updated files

+++ src/main/java/com/Medverse/Login/Repo/RegistrationRepo.java
@@ -1,3 +1,4 @@
+
 package com.Medverse.Login.Repo;
 
 import java.util.Optional;
@@ -8,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.Medverse.Login.Entity.RegistrationEntity;
 
 @Repository
-public interface RegistrationRepo extends JpaRepository<RegistrationEntity, Integer> {  // Use Integer as the type for the ID
+public interface RegistrationRepo extends JpaRepository<RegistrationEntity, Long> {  // Use Integer as the type for the ID
 
     Optional<RegistrationEntity> findByEmail(String email);  // Find user by email (case-sensitive by default)
 //    forget password...................................................................................
@@ -17,3 +18,4 @@ public interface RegistrationRepo extends JpaRepository<RegistrationEntity, Inte
 //..........................................................................................................
 
 }
+
